+++ lib/class-wp-theme-json-gutenberg.php
@@ -286,6 +286,7 @@ class WP_Theme_JSON_Gutenberg {
 		'--wp--style--root--padding-left'   => array( 'spacing', 'padding', 'left' ),
 		'text-decoration'                   => array( 'typography', 'textDecoration' ),
 		'text-transform'                    => array( 'typography', 'textTransform' ),
+		'word-break'                        => array( 'typography', 'wordBreak' ),
 		'filter'                            => array( 'filter', 'duotone' ),
 		'box-shadow'                        => array( 'shadow' ),
 		'writing-mode'                      => array( 'typography', 'writingMode' ),
@@ -458,6 +459,7 @@ class WP_Theme_JSON_Gutenberg {
 			'textColumns'      => null,
 			'textDecoration'   => null,
 			'textTransform'    => null,
+			'wordBreak'        => null,
 			'writingMode'      => null,
 		),
 	);
@@ -555,6 +557,7 @@ class WP_Theme_JSON_Gutenberg {
 			'textColumns'    => null,
 			'textDecoration' => null,
 			'textTransform'  => null,
+			'wordBreak'      => null,
 			'writingMode'    => null,
 		),
 		'css'        => null,